Revert "Lounge 페이지"
