+++ src/app/page.tsx
@@ -1,7 +1,3 @@
 export default async function Home() {
-  return (
-    <>
-
-    </>
-  );
+  return <></>;
 }
