+++ src/components/common/CardList/PbCardList.tsx
@@ -1,0 +1,16 @@
+"use client";
+import React from "react";
+import PbCardItem from "@/components/common/CardItem/PbCardItem";
+
+function PbCardList(props: any) {
+  
+  return (
+    <ul>
+      {props.map((item: any) => (
+        <PbCardItem key={item.id} item={item}/>
+      ))}
+    </ul>
+  );
+}
+
+export default PbCardList;
