+++ src/mocks/hyeon17/Common/pbContent.json
@@ -15,7 +15,7 @@
       },
       {
         "id": 2,
-        "title": "제목입니다2",
+        "title": "제목입니다1",
         "pbName": "김피비",
         "companyLogo": "logo.png",
         "career": 10,
@@ -25,7 +25,7 @@
       },
       {
         "id": 3,
-        "title": "제목입니다3",
+        "title": "제목입니다1",
         "pbName": "김피비",
         "companyLogo": "logo.png",
         "career": 10,
@@ -35,7 +35,7 @@
       },
       {
         "id": 4,
-        "title": "제목입니다4",
+        "title": "제목입니다1",
         "pbName": "김피비",
         "companyLogo": "logo.png",
         "career": 10,
@@ -45,7 +45,7 @@
       },
       {
         "id": 5,
-        "title": "제목입니다5",
+        "title": "제목입니다1",
         "pbName": "김피비",
         "companyLogo": "logo.png",
         "career": 10,
